+++ app.js
@@ -292,7 +292,6 @@ const readFileAsync = () => {
           let fname = station_ref.concat('_', sensor_ref)
 
    // console.log(i + '---'+ fname);
-          var fs = require('fs')
           var file = fs.createWriteStream('./public/data/Environment/water_levels/' + fname + '.csv')
           var http = require('http')
    // http://waterlevel.ie/data/month/25017_0001.csv
